pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-peer-review-env"
version = "0.1.0"
description = "OpenEnv benchmark for learning peer review from OpenReview/NeurIPS snapshots"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "openai>=2.30.0",
    "pydantic>=2.12.0",
    "requests>=2.33.0",
    "uvicorn>=0.44.0",
    "openreview-py>=2.0.1",
    "pymupdf>=1.27.2.2",
    "pymupdf4llm>=1.27.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
]

[project.scripts]
server = "peer_review_env.server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
peer_review_env = ["data_snapshot.json"]

[tool.setuptools.packages.find]
where = ["."]
include = ["peer_review_env*"]
